# maas-service/docs/swagger.yaml
definitions:
...
  model.Classifier:
    properties:
      name:
        type: string
      namespace:
        type: string
      tenantId:
        type: string
    required:
    - name
    - namespace
    type: object
...
  model.KafkaCredentials:
    properties:
      clientCert:
        type: string
      clientKey:
        type: string
      password:
        type: string
      type:
        type: string
      username:
        type: string
    type: object
...
  model.SyncTenantsResp:
    properties:
      tenant:
        $ref: '#/definitions/model.Tenant'
      topics:
        items:
          $ref: '#/definitions/model.TopicRegistrationRespDto'
        type: array
    type: object
  model.Tenant:
    properties:
      externalId:
        example: "101"
        type: string
      namespace:
        type: string
      tenantPresentation:
        additionalProperties: true
        type: object
    type: object
...
  model.TopicRegistrationRespDto:
    properties:
      actualSettings:
        $ref: '#/definitions/model.TopicSettings'
      addresses:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      caCert:
        example: ' '
        type: string
      classifier:
        $ref: '#/definitions/model.Classifier'
      credential:
        additionalProperties:
          items:
            $ref: '#/definitions/model.KafkaCredentials'
          type: array
        type: object
      externallyManaged:
        type: boolean
      instance:
        example: localkafka
        type: string
      name:
        example: local.sampleTopic2.430d243ba9c74f25bf05f492a52804fd
        type: string
      namespace:
        example: local
        type: string
      requestedSettings:
        $ref: '#/definitions/model.TopicSettings'
      template:
        example: "2"
        type: string
      versioned:
        example: false
        type: boolean
    type: object
...
  model.TopicSettings:
    properties:
      configs:
        additionalProperties:
          type: string
        type: object
      minNumPartitions:
        example: 1
        type: integer
      numPartitions:
        example: 1
        type: integer
      replicaAssignment:
        additionalProperties:
          items:
            type: integer
          type: array
        type: object
      replicationFactor:
        example: 1
        type: integer
      versioned:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: API for Maas Service.
  title: Maas Service API
paths:
...
  /api/v1/synchronize-tenants:
    post:
      description: Sync Tenants
      operationId: SyncTenants
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/model.Tenant'
          type: array
      - description: X-Origin-Namespace
        in: header
        name: X-Origin-Namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SyncTenantsResp'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Sync Tenants using Agent Role
      tags:
      - V1
  /api/v1/tenants:
    get:
      description: Get Tenants By Namespace
      operationId: GetTenantsByNamespace
      parameters:
      - description: X-Origin-Namespace
        in: header
        name: X-Origin-Namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Tenant'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BasicAuth:
        - agent
      summary: Get Tenants By Namespace using Agent Role
      tags:
      - V1
...
produces:
- application/json
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"